refactor: move vitePluginServerAssets
