+++ packages/react-server/src/plugin/index.ts
@@ -1,11 +1,8 @@
-import fs from "node:fs";
-import path from "node:path";
 import { fileURLToPath } from "node:url";
-import { createDebug, tinyassert, typedBoolean } from "@hiogawa/utils";
+import { createDebug, tinyassert } from "@hiogawa/utils";
 import {
   type ConfigEnv,
   type InlineConfig,
-  type Manifest,
   type Plugin,
   type PluginOption,
   type ViteDevServer,
@@ -13,6 +10,10 @@ import {
   createLogger,
   createServer,
 } from "vite";
+import {
+  SERVER_CSS_PROXY,
+  vitePluginServerAssets,
+} from "../features/assets/plugin";
 import {
   vitePluginClientUseServer,
   vitePluginServerUseServer,
@@ -22,13 +23,11 @@ import {
   vitePluginServerUseClient,
 } from "../features/use-client/plugin";
 import { $__global } from "../lib/global";
-import { collectStyle, collectStyleUrls } from "./css";
 import {
   ENTRY_CLIENT,
   ENTRY_CLIENT_WRAPPER,
   ENTRY_REACT_SERVER,
   ENTRY_REACT_SERVER_WRAPPER,
-  type SsrAssetsType,
   createVirtualPlugin,
   vitePluginSilenceDirectiveBuildWarning,
 } from "./utils";
@@ -50,6 +49,7 @@ const RUNTIME_REACT_SERVER_PATH = fileURLToPath(
 // convenient singleton to share states
 export type { ReactServerManager };
 
+// TODO: rename to PluginStateManager
 class ReactServerManager {
   parentServer?: ViteDevServer;
 
@@ -318,77 +318,13 @@ export function vitePluginReactServer(options?: {
       ssrRuntimePath: RUNTIME_SERVER_PATH,
     }),
     ...vitePluginClientUseClient({ manager }),
-    createVirtualPlugin("ssr-assets", async () => {
-      // dev
-      if (!manager.buildType) {
-        // extract <head> injected by plugins
-        let { head } = await getIndexHtmlTransform($__global.dev.server);
-
-        // expose raw dynamic `import` which doesn't go through vite's transform
-        // since it would inject `<id>?import` and cause dual packages when
-        // client code is both imported at the boundary (as `<id>?import`)
-        // and not at the boundary (as `<id>`).
-        head += `<script>globalThis.__raw_import = (id) => import(id)</script>\n`;
-
-        // serve dev css as ?direct so that ssr html won't get too huge.
-        // also remove style on first hot update.
-        head += `\
-          <link
-            data-ssr-dev-css
-            rel="stylesheet"
-            href="/@id/__x00__virtual:dev-ssr-css.css?direct"
-          />
-          <script type="module">
-            import { createHotContext } from "/@vite/client";
-            const hot = createHotContext("hot-data-ssr-dev-css");
-            hot.on("vite:afterUpdate", () => {
-              document
-                .querySelectorAll("[data-ssr-dev-css]")
-                .forEach(node => node.remove());
-            });
-          </script>
-        `;
-        const result: SsrAssetsType = {
-          bootstrapModules: [`/@id/__x00__${ENTRY_CLIENT_WRAPPER}`],
-          head,
-        };
-        return `export default ${JSON.stringify(result)}`;
-      }
-
-      // build
-      if (manager.buildType === "ssr") {
-        const manifest: Manifest = JSON.parse(
-          await fs.promises.readFile(
-            "dist/client/.vite/manifest.json",
-            "utf-8",
-          ),
-        );
-        const entry = manifest[ENTRY_CLIENT_WRAPPER];
-        tinyassert(entry);
-        const css = entry.css ?? [];
-        const js =
-          entry.dynamicImports
-            ?.map((k) => manifest[k]?.file)
-            .filter(typedBoolean) ?? [];
-        const head = [
-          ...css.map((href) => `<link rel="stylesheet" href="/${href}" />`),
-          ...js.map((href) => `<link rel="modulepreload" href="/${href}" />`),
-        ].join("\n");
-        const result: SsrAssetsType = {
-          bootstrapModules: [`/${entry.file}`],
-          head,
-        };
-        return `export default ${JSON.stringify(result)}`;
-      }
-
-      tinyassert(false);
-    }),
+    ...vitePluginServerAssets({ manager }),
     createVirtualPlugin(ENTRY_CLIENT_WRAPPER.slice("virtual:".length), () => {
       // dev
       if (!manager.buildType) {
         // wrapper entry to ensure client entry runs after vite/react inititialization
         return /* js */ `
-          import "virtual:react-server-css.js";
+          import "${SERVER_CSS_PROXY}";
           for (let i = 0; !window.__vite_plugin_react_preamble_installed__; i++) {
             await new Promise(resolve => setTimeout(resolve, 10 * (2 ** i)));
           }
@@ -399,58 +335,12 @@ export function vitePluginReactServer(options?: {
       if (manager.buildType === "client") {
         // import "runtime-client" for preload
         return /* js */ `
-          import "virtual:react-server-css.js";
+          import "${SERVER_CSS_PROXY}";
           import("@hiogawa/react-server/runtime-client");
           import "${ENTRY_CLIENT}";
         `;
       }
       tinyassert(false);
     }),
-    createVirtualPlugin("dev-ssr-css.css?direct", async () => {
-      tinyassert(!manager.buildType);
-      const styles = await Promise.all([
-        `/******* react-server ********/`,
-        collectStyle($__global.dev.reactServer, [ENTRY_REACT_SERVER]),
-        `/******* client **************/`,
-        collectStyle($__global.dev.server, [ENTRY_CLIENT]),
-      ]);
-      return styles.join("\n\n");
-    }),
-    createVirtualPlugin("react-server-css.js", async () => {
-      // virtual module proxy css imports from react server to client
-      // TODO: invalidate + full reload when add/remove css file?
-      if (!manager.buildType) {
-        const urls = await collectStyleUrls($__global.dev.reactServer, [
-          ENTRY_REACT_SERVER,
-        ]);
-        const code = urls.map((url) => `import "${url}";\n`).join("");
-        // ensure hmr boundary since css module doesn't have `import.meta.hot.accept`
-        return code + `if (import.meta.hot) { import.meta.hot.accept() }`;
-      }
-      if (manager.buildType === "client") {
-        // TODO: probe manifest to collect css?
-        const files = await fs.promises.readdir("./dist/rsc/assets", {
-          withFileTypes: true,
-        });
-        const code = files
-          .filter((f) => f.isFile() && f.name.endsWith(".css"))
-          .map((f) => path.join(f.path, f.name))
-          .map((f) => `import "/${f}";\n`)
-          .join("");
-        return code;
-      }
-      tinyassert(false);
-    }),
   ];
-}
-
-async function getIndexHtmlTransform(server: ViteDevServer) {
-  const html = await server.transformIndexHtml(
-    "/",
-    "<html><head></head></html>",
-  );
-  const match = html.match(/<head>(.*)<\/head>/s);
-  tinyassert(match && 1 in match);
-  const head = match[1];
-  return { head };
 }
